1. Create BasicAssetExtractor.cs and put ExtractAssets() with GetProcessInfo methods. ( code will be shared with VmAgent ) 2. Pass Global file extention parameters to BasicAssetExtractor. 3. Pass the instance of BasicAssetExtractor to MultiplayerServerManager
